requirements.txt
psycopg2-binary
reportlab
django-widget-tweaks
